Refactor components, move plugin-related files to plugin subdir
[#167306615]
