+++ src/components/authoring/glossary-resource-selector.scss
