+++ src/components/common/icons.scss
@@ -9,11 +9,11 @@
 
 @font-face {
   font-family: 'icomoon';
-  src:  url('../fonts/icomoon.eot?1g12qr');
-  src:  url('../fonts/icomoon.eot?1g12qr#iefix') format('embedded-opentype'),
-  url('../fonts/icomoon.ttf?1g12qr') format('truetype'),
-  url('../fonts/icomoon.woff?1g12qr') format('woff'),
-  url('../fonts/icomoon.svg?1g12qr#icomoon') format('svg');
+  src:  url('../../fonts/icomoon.eot?1g12qr');
+  src:  url('../../fonts/icomoon.eot?1g12qr#iefix') format('embedded-opentype'),
+  url('../../fonts/icomoon.ttf?1g12qr') format('truetype'),
+  url('../../fonts/icomoon.woff?1g12qr') format('woff'),
+  url('../../fonts/icomoon.svg?1g12qr#icomoon') format('svg');
   font-weight: normal;
   font-style: normal;
 }
