+++ src/components/plugin/definition.scss
