+++ src/components/plugin/glossary-popup.scss
