+++ src/components/plugin/glossary-sidebar.scss
