+++ src/components/plugin/plugin-app.scss
